Add Sensu handler if there is a field notification_handler
A notification handler can be configured in the sensu encoder and we can
overwrite it by setting a key 'notification_handler' in the alarm
definition and enabling notification. A noop handler that is doing
nothing can also be set and it will be used to manage cases where
alerting is enabled without notifications.

Change-Id: I1f38fb0aa6c27a3a8c2f582067f8c7af9b0c1a4b